Added optional format string to email subject line.
Consider this use case. Say you've got an item.title set to "Foo Bar Live". For Facebook sharing, you want to just use the title as is, but you'd like to add a bit of boilerplate to email subject lines…Subject: I'm going to 'Foo Bar Live'! This change makes that possible with a simple edit in Localizable.strings.