SHKShareMenu.m: fix one crash and one bug
If exclusions were set in the user defaults and a user subsequently
tried to change the exclusions again, the app would crash due to the
array not being mutable.
Additionally, removing exclusions was not always possible because
removeObjectIdenticalTo: was used instead of removeObject:. The formes
compares addresses which might be fine during the same session (did not
check this one completely), but after starting the app again the
addresses are very likely to differ.