Re-added accidentially replaced support for custom HTML body
We are setting a shorter text on item.text for short message sharers like Twitter. But for emails we were using the custom value "body" to set an elaborately designed long form email.

You apparently have removed this by accident, this re-adds this functionality. If a body custom value exists then this is used, otherwise item.text is used as before.